Declared reserved load order slots.
Adjusted CheatModule load order to match the reserved load order slots.
