+++ Framework/docs/modules_loadorder.md
@@ -1,0 +1,15 @@
+# CheatModule Load Order
+
+## Purpose of this file
+
+To declare a standard load order for CheatModules (a.k.a. Addons)
+
+## Reserved Load Order Slots
+
+- 0-19:     Console Command Only modules
+- 20-29:    Hotkey Only modules
+- 30-49:    Root Menu modules
+- 50-59:    Race Changer Support modules
+- 60-69:    Toggle Cheats SubMenu modules
+
+All other slots are free to use.
