+++ Modules/Configurable Infinite Stats Cheat/addons/61_CheatModule_InfiniteMainStats.rb
