+++ Modules/Hair Color Editor/addons/34_CheatModule_HairColorEdit.rb
