+++ Modules/Infinite Money Cheat/addons/63_CheatModule_InfiniteMoney.rb
