+++ Modules/Inventory Editor/addons/30_CheatModule_InvEdit.rb
