+++ Modules/Legacy Cheats/addons/49_CheatModule_Legacy.rb
