+++ Modules/Morality Editor/addons/36_CheatModule_MoralityEdit.rb
