+++ Modules/Sea Witch Awaken Skill/addons/51_CheatModule_DeepSkill.rb
