+++ Modules/Unequip Items/addons/20_CheatModule_UnequipItems.rb
